evaluation/alternativeClusteringApproaches/embedding/requirements.txt
joblib
backoff
numpy
pandas
umap-learn
hdbscan
streamlit
plotly
tqdm
python-dotenv
google.genai
google.api_core